Move `VersionInfo` to dedicated module
